% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abc2num}
\alias{abc2num}
\title{Convert letters to a number, e.g. c becomes 3}
\usage{
abc2num(s, base = 26)
}
\arguments{
\item{s}{The string to convert (lowercase letters)}

\item{base}{The number of letters to use (default 26 for a-z)}
}
\description{
Convert letters to a number, e.g. c becomes 3
}
