% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Add columns if they are missing from the data frame}
\usage{
add_missing_columns(df, cols, default = NA)
}
\arguments{
\item{df}{A data frame}

\item{cols}{A character vector with column names}

\item{default}{A value to set as default for missing columns}
}
\value{
A data frame with new columns if necessary
}
\description{
Add columns if they are missing from the data frame
}
\keyword{internal}
