% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_download}
\alias{api_download}
\title{Download a file from Epigraf}
\usage{
api_download(
  endpoint,
  params = c(),
  filename = NULL,
  filepath = NULL,
  overwrite = FALSE,
  database = NA
)
}
\arguments{
\item{endpoint}{The endpoint path.}

\item{params}{Query parameters.}

\item{filename}{A file name or a full path to the local file.}

\item{filepath}{A target folder path}

\item{overwrite}{Whether to overwrite existing files.}

\item{database}{The selected database.}
}
\value{
A list with error and data elements.
}
\description{
Download a file from Epigraf
}
