% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{api_fetch}
\alias{api_fetch}
\title{Fetch entity data such as articles, projects or properties}
\usage{
api_fetch(table, params = c(), db, maxpages = 1)
}
\arguments{
\item{table}{The table name (e.g. "articles")}

\item{params}{A named list of query params}

\item{db}{The database name}

\item{maxpages}{Maximum number of pages to request.
Set to 1 for non-paginated tables.}
}
\description{
Fetch entity data such as articles, projects or properties
}
