% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch}
\alias{api_patch}
\title{Patch data}
\usage{
api_patch(data, db, table = NA, type = NA, wide = T)
}
\arguments{
\item{data}{A dataframe with the column \code{id}.
Additional columns such as \code{norm_data} will be written to the entity.
The id must either be a valid IRI path (e.g. \code{properties/objecttypes/xxx})
or an id prefixed by the table name (e.g. \code{properties-12}).
Patching properties with prefixed ids requires a \code{type} column
that contains the property type.
If wide is set to TRUE (default), column names prefixed with table names are extracted.}

\item{db}{The database name.}

\item{table}{Optional: Check that the data only contains rows for a specific table.}

\item{type}{Optional: Check that the data only contains rows with a specific type.}

\item{wide}{Convert wide format to long format.
If TRUE, column names prefixed with "properties", "items", "sections", "articles"
and "projects" followed by a dot (e.g. \code{properties.id}, \code{properties.lemma})
will be extracted and patched as additional entities.}
}
\description{
Update entities in the database using the API.
Existing entities will be updated, missing entities will be created.
The function supports uploading all data related to articles:
articles, sections, items, links, footnotes, properties, projects, users, types.
The IRI path in the ID column of the dataframe must contain the specific table name.
}
