% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_post}
\alias{api_post}
\title{Post data to epigraf}
\usage{
api_post(endpoint, params = c(), payload = NULL, database = NA)
}
\arguments{
\item{endpoint}{The endpoint path}

\item{params}{Query parameters}

\item{payload}{The data posted to the endpoint}

\item{database}{The selected database}
}
\value{
void
}
\description{
Post data to epigraf
}
