% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_table}
\alias{api_table}
\title{Download tables}
\usage{
api_table(
  endpoint,
  params = c(),
  db = NA,
  maxpages = 1,
  compact = FALSE,
  silent = FALSE
)
}
\arguments{
\item{endpoint}{The endpoint path (e.g. "articles/index" or "articles/view/1")}

\item{params}{A named list of query params}

\item{db}{The database name.
Provide a character vector of dababase names to get and row bind data from multiple databases.
In this case, the compact parameter is automatically set to TRUE. Thus, a database name column is added.}

\item{maxpages}{Maximum number of pages to request.
Set to 1 for non-paginated tables.}

\item{compact}{Whether to rename type columns to \code{type} and to add a \code{table} and a \code{database} column.}

\item{silent}{Whether to output status messages}
}
\description{
Fetches tables such as articles, projects or properties
}
\details{
TODO: silent problem message (false positive, results from the last column being empty from case to case)
TODO: add progress bar
}
