% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_upload}
\alias{api_upload}
\title{Upload file to epigraf}
\usage{
api_upload(
  endpoint,
  params = c(),
  filepath = NULL,
  mimetype = NULL,
  database = NA
)
}
\arguments{
\item{endpoint}{The endpoint path}

\item{params}{Query parameters}

\item{filepath}{A full path to the local file}

\item{mimetype}{The mime type fof the file. Will be guessed if empty.}

\item{database}{The selected database}
}
\value{
void
}
\description{
Upload file to epigraf
}
