% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{craft_items}
\alias{craft_items}
\title{Create RAM rows for item data}
\usage{
craft_items(df, cols = c(), fill = c())
}
\arguments{
\item{df}{The source data frame}

\item{cols}{The mapping of source columns to RAM columns}

\item{fill}{Fixed values}
}
\value{
A RAM-enhanced data frame
}
\description{
Create RAM rows for item data
}
