% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{craft_properties}
\alias{craft_properties}
\title{Create RAM rows for property data}
\usage{
craft_properties(df, cols, fill)
}
\arguments{
\item{df}{The source data frame}

\item{cols}{The mapping of source columns to RAM columns}

\item{fill}{Fixed values}
}
\value{
RAM rows
}
\description{
Create RAM rows for property data
}
\keyword{internal}
