% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_condition}
\alias{db_condition}
\title{Construct filter conditions for the db_table() function}
\usage{
db_condition(table = NA, field, value)
}
\arguments{
\item{table}{Table name. Leave empty to omit the table prefix.}

\item{field}{Field name.}

\item{value}{A single value, a list of characters or a list of integers.}
}
\description{
Construct filter conditions for the db_table() function
}
