% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_databases}
\alias{db_databases}
\title{Get list of all databases}
\usage{
db_databases(epi = TRUE)
}
\arguments{
\item{epi}{Only keep databases with the epi-prefix and join meta data from the main database}
}
\description{
Get list of all databases
}
