% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_table}
\alias{db_table}
\title{Get data from a database table}
\usage{
db_table(table, cond = list(), deleted = FALSE, compact = FALSE, db)
}
\arguments{
\item{table}{Table name.}

\item{cond}{Either a named list of conditions, or a full condition as character, or a character vector of conditions, e.g. \code{c("id = 10")}.}

\item{deleted}{Deleted records are skipped by default. Set to TRUE, to get all records.}

\item{compact}{Whether to rename type columns to \code{type} and to add a \code{table} and a \code{database} column.}

\item{db}{A connection object (object) or the database name (character).
Provide a character vector of dababase names to get and row bind data from multiple databases.
In this case, the compact parameter is automatically set to TRUE. Thus, a database name column is added.}
}
\description{
Get data from a database table
}
