% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{df_to_ram}
\alias{df_to_ram}
\title{Map a data frame to the Relational Article Model}
\usage{
df_to_ram(
  df,
  project.cols = c(),
  project.fill = c(),
  article.cols = c(),
  article.fill = c(),
  section.cols = c(),
  section.fill = c(),
  property.cols = c(),
  property.fill = c(),
  item.cols = c(),
  item.fill = c(),
  compile = FALSE
)
}
\arguments{
\item{df}{A data frame with the source data.}

\item{project.cols}{A named list. Names are RAM columns, values source columns.}

\item{project.fill}{A named list with fixed values for the rows.}

\item{article.cols}{A named list. Names are RAM columns, values source columns.}

\item{article.fill}{A named list with fixed values for the rows.}

\item{section.cols}{A named list. Names are RAM columns, values source columns.}

\item{section.fill}{A named list with fixed values for the rows.}

\item{property.cols}{A named list. Names are RAM columns, values source columns.}

\item{property.fill}{A named list with fixed values for the rows.}

\item{item.cols}{A named list. Names are RAM columns, values source columns.}

\item{item.fill}{A named list with fixed values for the rows.}

\item{compile}{Whether to return compiled rows or store them in the attributes of the input data frame.}
}
\value{
A data frame. When compile is TRUE, ready to be imported into Epigraf.
Otherwise with rows in the attributes of the data frame.
Call \code{ram_compile()} to get the import data.
}
\description{
Map a data frame to the Relational Article Model
}
\examples{
library(tibble)
library(eprgraf)

df <- tribble(
  ~case, ~title, ~text,
  1, "Case 01", "Happy New Year!",
  2, "Case 02", "Happy Easter!",
  3, "Case 03", "Happy Birthday!"
)

df |>
 df_to_ram(
   project.fill = c("fragment" = "Example"),
   item.cols = c("content" = "text"),
   article.cols = c("id" = "case", "signature" = "case", "name" = "title"),
   compile = T
 )

}
