% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill_items}
\alias{distill_items}
\title{Get articles (including selected item values)}
\usage{
distill_items(
  df,
  type = NULL,
  cols = c(),
  property.cols = c(),
  article.cols = c()
)
}
\arguments{
\item{df}{A RAM data frame}

\item{type}{Item types to filter}

\item{cols}{Cols returned from the items}

\item{property.cols}{Property columns joined to the items}

\item{article.cols}{Article columns joined to the items. Not implemented yet.}
}
\value{
A tibble with items
}
\description{
TODO: Implement article.cols parameter.
}
\keyword{internal}
