% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.api_progress}
\alias{.api_progress}
\title{Show a progress bar based on job progress fields}
\usage{
.api_progress(pb, body)
}
\arguments{
\item{pb}{The current progress bar isntance or NULL}

\item{body}{The job body}
}
\value{
The progress bar instance
}
\description{
Show a progress bar based on job progress fields
}
\keyword{internal}
