% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_empty_columns}
\alias{drop_empty_columns}
\title{Remove empty columns}
\usage{
drop_empty_columns(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame without columns that only contain NA values
}
\description{
Remove empty columns
}
