% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{epi_iri_parent}
\alias{epi_iri_parent}
\title{Get the IRI fragment of an IRI path}
\usage{
epi_iri_parent(id = NULL, prefix = "~")
}
\arguments{
\item{id}{An IRI path}

\item{prefix}{A prefix added to the IRI fragment, if the ID is not NULL.}
}
\value{
The IRI fragment
}
\description{
Get the IRI fragment of an IRI path
}
