% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{epi_is_irifragment}
\alias{epi_is_irifragment}
\title{Check whether the provided vector contains only valid IRI fragments}
\usage{
epi_is_irifragment(value)
}
\arguments{
\item{value}{The vector that will be checked.}
}
\value{
A vector of the same length as value, containg TRUE or FALSE.
}
\description{
Check whether the provided vector contains only valid IRI fragments
}
