% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigraf-package.R
\docType{package}
\name{eprgraf-package}
\alias{eprgraf}
\alias{eprgraf-package}
\title{eprgraf: Fetch from and push data to Epigraf}
\description{
Provides functions for interacting with Epigraf.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datavana/eprgraf}
}

}
\author{
\strong{Maintainer}: Jakob Jünger \email{jakob.juenger@uni-muenster.de} (\href{https://orcid.org/0000-0003-1860-6695}{ORCID})

}
\keyword{internal}
