% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{extract_segment}
\alias{extract_segment}
\title{Function to extract segments based on ID attribute}
\usage{
extract_segment(xml, tagid)
}
\arguments{
\item{xml}{Character value containing XML text}

\item{tagid}{Character value containing the tag ID}
}
\value{
A character value containing only the text of elements with the tag ID
}
\description{
Function to extract segments based on ID attribute
}
