% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{extract_untagged}
\alias{extract_untagged}
\title{Function to extract non-tagged text}
\usage{
extract_untagged(xml)
}
\arguments{
\item{xml}{The XML as character value.}
}
\value{
A character value where all text contained in tags was stripped.
}
\description{
Function to extract non-tagged text
}
\keyword{internal}
