% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_extension}
\alias{get_extension}
\title{Get file extension from URL path component}
\usage{
get_extension(path)
}
\arguments{
\item{path}{A file path}
}
\value{
The file extension, without dot
}
\description{
Get file extension from URL path component
}
\keyword{internal}
