% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isLocalServer}
\alias{isLocalServer}
\title{Check whether the URL is on a local server}
\usage{
isLocalServer(server)
}
\arguments{
\item{server}{The server URL}
}
\description{
Check whether the URL is on a local server
}
\keyword{internal}
