% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_path}
\alias{join_path}
\title{Join folder and filename}
\usage{
join_path(filename, filepath = NULL)
}
\arguments{
\item{filename}{File name}

\item{filepath}{Target folder or NULL}
}
\value{
Joined path.
}
\description{
Join folder and filename
}
\keyword{internal}
