% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_vectors}
\alias{merge_vectors}
\title{Merge vectors}
\usage{
merge_vectors(values, default)
}
\arguments{
\item{values}{A named vector}

\item{default}{A second named vector}
}
\value{
The first vector containing all additional named values from the second vector
}
\description{
Merge vectors
}
