% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_cols_to_end}
\alias{move_cols_to_end}
\title{Shift selected columns to the end}
\usage{
move_cols_to_end(df, cols)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{A vector of column names}
}
\value{
A dataframe ending with the selected cols if they are present
}
\description{
Shift selected columns to the end
}
