% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_cols_to_front}
\alias{move_cols_to_front}
\title{Shift selected columns to the front}
\usage{
move_cols_to_front(df, cols)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{A vector of column names}
}
\value{
A dataframe starting with the selected cols if they are present
}
\description{
Shift selected columns to the front
}
