% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.omit.list}
\alias{na.omit.list}
\title{Remove NA entries from a list}
\usage{
\method{na.omit}{list}(object, ...)
}
\arguments{
\item{...}{Not used, for compatibility with na.omit()}
}
\value{
A cleaned list
}
\description{
Remove NA entries from a list
}
\keyword{internal}
