% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_add}
\alias{ram_add}
\title{Add a row to the rows in the epi attribute}
\usage{
ram_add(df, rows, skip = FALSE)
}
\arguments{
\item{df}{A tibble}

\item{rows}{Crafted rows}

\item{skip}{Whether to update the record or only use it as reference}
}
\value{
Epigraf tibble
}
\description{
Add a row to the rows in the epi attribute
}
