% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_clear}
\alias{ram_clear}
\title{Remove all RAM rows from a tibble}
\usage{
ram_clear(df)
}
\arguments{
\item{df}{A tibble with RAM rows in the epi attribute}
}
\value{
Epigraf tibble without RAM rows
}
\description{
Remove all RAM rows from a tibble
}
\examples{
library(tibble)
library(eprgraf)

# Example data
df <- tribble(
  ~case, ~title, ~text,
  1, "Case 01", "Happy New Year!",
  2, "Case 02", "Happy Easter!",
  3, "Case 03", "Happy Birthday!"
)

# Add RAM rows
df <- df |>
 df_to_ram(
   item.cols = c("content" = "text"),
   article.cols = c("id" = "case", "signature" = "case", "name" = "title")
 )

ram_compile(df)

# Remove RAM rows
df <-ram_clear(df)
ram_compile(df)

}
