% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_compile}
\alias{ram_compile}
\title{Compile a crafted table with RAM rows ready to patch into the database}
\usage{
ram_compile(df)
}
\arguments{
\item{df}{An Epigraf tibble with RAM rows in the epi attribute}
}
\value{
The RAM rows data frame

A tibble with RAM rows
}
\description{
RAM rows can be added with \code{ram_add()}.
}
\examples{
library(tibble)
library(rpigraf)

# Example data
df <- tribble(
  ~case, ~title, ~text,
  1, "Case 01", "Happy New Year!",
  2, "Case 02", "Happy Easter!",
  3, "Case 03", "Happy Birthday!"
)

# Add RAM rows
df <- df |>
 df_to_ram(
   item.cols = c("content" = "text"),
   article.cols = c("id" = "case", "signature" = "case", "name" = "title")
 )

ram_compile(df)

}
