% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_add_ancestor}
\alias{tree_add_ancestor}
\title{Add ancestor id and path from a specific level to all children}
\usage{
tree_add_ancestor(data, level = 0, col_id, col_parent_id, col_path)
}
\arguments{
\item{data}{A data frame with properties}

\item{level}{The target level}

\item{col_id}{The column holding property ids}

\item{col_parent_id}{The column holding property parent ids}

\item{col_path}{The column holding a value (mostly a path or lemma) that will be added in addition to the id.}
}
\value{
Data frame with the two columns ancestor_id and ancestor_path added
}
\description{
Add ancestor id and path from a specific level to all children
}
