% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unescape_html}
\alias{unescape_html}
\title{Remove HTML entities}
\usage{
unescape_html(str)
}
\arguments{
\item{str}{The input character value}
}
\value{
The output character value
}
\description{
Remove HTML entities
}
\keyword{internal}
