% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{api_fetch_entity}
\alias{api_fetch_entity}
\title{Fetch entities such as single articles, projects or properties}
\usage{
api_fetch_entity(ids, params = c(), db = NULL, silent = FALSE)
}
\arguments{
\item{ids}{A character vector with IDs as returned by api_fetch_table, e.g. articles-1.
Alternatively, provide a dataframe containg the IDs in the id-column.
So you can chain fetch_articles() and api_fetch_entity()}

\item{params}{A named list of query params}

\item{db}{The database name. Leave empty when providing a dataframe produced by api_fetch_table().
In this case, the database name will be extracted from the dataframe.}

\item{silent}{Whether to output a progress bar}
}
\description{
Returns all data belonging to the entity identified by ID.
The procedure corresponds to calling the view action in the Epigraf interface.
}
