% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{craft_projects}
\alias{craft_projects}
\title{Create RAM rows for project data}
\usage{
craft_projects(df, mapping, fill)
}
\arguments{
\item{df}{The source data frame}

\item{mapping}{The mapping of source columns to RAM columns}

\item{fill}{Fixed values}
}
\value{
RAM rows
}
\description{
Create RAM rows for project data
}
\keyword{internal}
