% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{db_fetch}
\alias{db_fetch}
\title{Fetch entity data such as articles, projects or properties using direct database access}
\usage{
db_fetch(table, params = list(), db = NA)
}
\arguments{
\item{table}{The table name (e.g. "articles").}

\item{params}{A named list of query conditions, passed to db_table.}

\item{db}{The database name (character).
Provide a character vector of dababase names to get and row bind data from multiple databases.}
}
\description{
Returns all data belonging to all entities matched by the params.
}
