% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{df_to_ram}
\alias{df_to_ram}
\title{Map a data frame to the Relational Article Model (RAM)}
\usage{
df_to_ram(
  df,
  project.cols = c(),
  project.fill = c(),
  article.cols = c(),
  article.fill = c(),
  section.cols = c(),
  section.fill = c(),
  property.cols = c(),
  property.fill = c(),
  item.cols = c(),
  item.fill = c(),
  compile = FALSE
)
}
\arguments{
\item{df}{A data frame with the source data.}

\item{project.cols}{A named list. Names are RAM columns, values source columns.}

\item{project.fill}{A named list with fixed values for the rows.}

\item{article.cols}{A named list. Names are RAM columns, values source columns.}

\item{article.fill}{A named list with fixed values for the rows.}

\item{section.cols}{A named list. Names are RAM columns, values source columns.}

\item{section.fill}{A named list with fixed values for the rows.}

\item{property.cols}{A named list. Names are RAM columns, values source columns.}

\item{property.fill}{A named list with fixed values for the rows.}

\item{item.cols}{A named list. Names are RAM columns, values source columns.}

\item{item.fill}{A named list with fixed values for the rows.}

\item{compile}{Whether to return the compiled RAM rows or a RAM enhanced data frame.}
}
\value{
When compile is TRUE: a RAM data frame, ready for patching into the Epigraf database.
When compile is FALSE: a RAM-enhanced data frame.
In RAM-enhanced data frames, the RAM rows are stored in the epi attribute.
Call \code{ram_compile()} to get a RAM data frame from a RAM-enhanced data frame.
}
\description{
Calls craft methods for projects, articles, sections, items and properties.
Alternatively, you can use the craft methods directly.
}
\examples{
library(tibble)
library(rpigraf)

df <- tribble(
  ~case, ~title, ~text,
  1, "Case 01", "Happy New Year!",
  2, "Case 02", "Happy Easter!",
  3, "Case 03", "Happy Birthday!"
)

df |>
 df_to_ram(
   project.fill = c("fragment" = "Example"),
   item.cols = c("content" = "text"),
   article.cols = c("id" = "case", "signature" = "case", "name" = "title"),
   compile = T
 )

}
