% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill_articles}
\alias{distill_articles}
\title{Get articles including a selected item value}
\usage{
distill_articles(
  df,
  cols = c(),
  item.type = NULL,
  item.cols = c(),
  property.cols = c()
)
}
\arguments{
\item{df}{A RAM data frame}

\item{item.type}{Item types to join}

\item{item.cols}{Cols to join from the items}
}
\value{
A tibble with cases
}
\description{
Get articles including a selected item value
}
