% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill_articles}
\alias{distill_articles}
\title{Get articles}
\usage{
distill_articles(
  df,
  cols = c(),
  section.type = NULL,
  section.cols = c(),
  item.type = NULL,
  item.cols = c(),
  property.cols = c()
)
}
\arguments{
\item{df}{A RAM data frame.}

\item{cols}{Article columns.}

\item{section.type}{Section types to join.
The result contains only items within sections of the given type.
Set to NULL to get all items.}

\item{section.cols}{Cols to join from the sections.}

\item{item.type}{Item types to join.}

\item{item.cols}{Cols to join from the items.}

\item{property.cols}{Cols to join from the property.}
}
\value{
A tibble with articles.
}
\description{
Section, item, and property data can be joined.
}
