% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill_links}
\alias{distill_links}
\title{Get annotations for the articles}
\usage{
distill_links(
  df,
  items.type = NULL,
  properties.type = NULL,
  cols = c("path", "segment"),
  article.cols = c(),
  level = 0
)
}
\arguments{
\item{df}{A RAM data frame.}

\item{items.type}{The type of items with annotations.}

\item{properties.type}{Keep only links that target the given property type.}

\item{article.cols}{A list of article columns to join.}

\item{level}{The aggregation level, beginning with 0. Set to NULL to get the lowest level.}
}
\value{
A tibble containing annotations.
}
\description{
Get annotations for the articles
}
\keyword{internal}
