% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill_properties}
\alias{distill_properties}
\title{Get the property tree by type}
\usage{
distill_properties(df, type = NULL, cols = c(), annos = FALSE)
}
\arguments{
\item{df}{A RAM data frame}
}
\value{
A tibble containing the code tree
}
\description{
Get the property tree by type
}
