% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill_properties}
\alias{distill_properties}
\title{Get the property tree (including annotations)}
\usage{
distill_properties(df, type = NULL, cols = c(), annos = FALSE)
}
\arguments{
\item{df}{A RAM data frame.}

\item{type}{The property type.}

\item{cols}{The property columns.}

\item{annos}{Whether to distill annotations.}
}
\value{
A tibble containing the properties tree.
}
\description{
Get the property tree (including annotations)
}
