% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.api_request}
\alias{.api_request}
\title{Send request to epigraf}
\usage{
.api_request(
  endpoint,
  params = c(),
  payload = NULL,
  database = NA,
  method = httr::POST
)
}
\arguments{
\item{endpoint}{The endpoint supporting job creation}

\item{params}{Query parameters}

\item{payload}{The data posted to the endpoint}

\item{database}{The selected database}

\item{method}{One of the httr functions (httr::POST, httr::DELETE)}
}
\value{
void
}
\description{
Send request to epigraf
}
\keyword{internal}
