% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_clean_irifragment}
\alias{epi_clean_irifragment}
\title{Create a clean IRI fragment}
\usage{
epi_clean_irifragment(fragment)
}
\arguments{
\item{fragment}{The dirty IRI fragment that will be cleaned}
}
\value{
A clean IRI fragment
}
\description{
Replaces all non alphanumeric characters by hyphens
and converts the input to lowercase
}
