% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_extract_long}
\alias{epi_extract_long}
\title{Get RAM rows by table name}
\usage{
epi_extract_long(df, table, type = NULL, prefix = TRUE)
}
\arguments{
\item{df}{A RAM data frame}

\item{table}{The table name}

\item{type}{Filter by type}

\item{prefix}{Whether to prefix the columns with the table name}
}
\value{
A data frame with the filtered rows and columns prefixed with the table name
}
\description{
Get RAM rows by table name
}
