% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_wide_to_long}
\alias{epi_wide_to_long}
\title{Convert wide to long format}
\usage{
epi_wide_to_long(data)
}
\arguments{
\item{data}{A dataframe with the column id containing a valid IRI path.
If additional column names are prefixed with "properties", "items", "sections",
"articles" and "projects" followed by a dot (e.g. "properties.id",
"properties.lemma"), they will be extracted and stacked to the dataframe.}
}
\value{
A dataframe with all input rows and the nested entities stacked.
}
\description{
Convert wide to long format
}
